requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
langchain==0.1.0
langchain-groq==0.0.1
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
jinja2==3.1.2 